# gobackend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "OpenUpload API",
    "description": "API for managing file uploads, projects, and API keys",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://open-upload-api.gabrielmalek.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Firebase ID token"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for programmatic access"
      }
    },
    "schemas": {
...
      "DashboardStats": {
        "type": "object",
        "properties": {
          "total_storage": {
            "type": "integer",
            "format": "int64"
          },
          "total_storage_limit": {
            "type": "integer",
            "format": "int64"
          },
          "total_files": {
            "type": "integer",
            "format": "int64"
          },
          "total_api_requests": {
            "type": "integer",
            "format": "int64"
          },
          "api_requests_change": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "UsageStats": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "api_calls": {
            "type": "integer",
            "format": "int64"
          },
          "avg_response_time": {
            "type": "number",
            "format": "float"
          },
          "success_rate": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "ApiUsage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "endpoint": {
            "type": "string"
          },
          "response_time": {
            "type": "number",
            "format": "float"
          },
          "status_code": {
            "type": "integer"
          },
          "user_firebase_uid": {
            "type": "string"
          },
          "project_id": {
            "type": "integer"
          },
          "api_key_id": {
            "type": "integer"
          }
        }
      },
      "BucketStats": {
        "type": "object",
        "properties": {
          "total_size": {
            "type": "integer",
            "format": "int64",
            "description": "Total size in bytes"
          },
          "object_count": {
            "type": "integer",
            "format": "int64",
            "description": "Number of objects"
          }
        }
      },
      "StorageStats": {
        "type": "object",
        "properties": {
          "database_storage": {
            "type": "integer",
            "format": "int64",
            "description": "Storage tracked in database (user's files)"
          },
          "minio_storage": {
            "type": "integer",
            "format": "int64",
            "description": "Actual storage used in MinIO bucket (all users)"
          },
          "minio_objects": {
            "type": "integer",
            "format": "int64",
            "description": "Number of objects in MinIO bucket"
          },
          "storage_limit": {
            "type": "integer",
            "format": "int64",
            "description": "User storage limit in bytes"
          },
          "minio_stats": {
            "$ref": "#/components/schemas/BucketStats",
            "description": "Detailed MinIO bucket statistics"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    }
  },
  "paths": {
...
    "/usage/dashboard-stats": {
      "get": {
        "summary": "Get dashboard statistics",
        "description": "Get overall usage statistics for the current user",
        "operationId": "getDashboardStats",
        "tags": ["Usage"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardStats"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/usage/storage": {
      "get": {
        "summary": "Get storage statistics",
        "description": "Get detailed storage statistics including database-tracked storage and actual MinIO bucket usage",
        "operationId": "getStorageStats",
        "tags": ["Usage"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Storage statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageStats"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/usage": {
      "get": {
        "summary": "Get usage statistics",
        "description": "Get usage statistics grouped by date, optionally filtered by project and date range",
        "operationId": "getUsageStats",
        "tags": ["Usage"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "description": "Filter by project ID",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "description": "Start date (YYYY-MM-DD)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "description": "End date (YYYY-MM-DD)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usage statistics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UsageStats"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/usage/details": {
      "get": {
        "summary": "Get usage details",
        "description": "Get detailed API usage records, optionally filtered by project, API key, and date range",
        "operationId": "getUsageDetails",
        "tags": ["Usage"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "description": "Filter by project ID",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "api_key_id",
            "in": "query",
            "description": "Filter by API key ID",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "description": "Start date (YYYY-MM-DD)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "description": "End date (YYYY-MM-DD)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of records to return",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usage details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApiUsage"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}